end-migrate-storage/requirements.txt
flask
google-cloud-datastore
google-cloud-storage